Backend/requirements.txt
# DarkHook Defense Backend Requirements

# Web Framework
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.12

# Database
pymongo>=4.10.0

# ML Models
scikit-learn>=1.6.0
xgboost>=2.1.0
numpy>=2.0.0

# PDF Analysis
PyMuPDF>=1.25.0

# OCR
pytesseract>=0.3.10
Pillow>=11.0.0

# Office Files
python-docx>=1.1.0
openpyxl>=3.1.0
oletools>=0.60.2

# QR Code Detection
pyzbar>=0.1.9

# Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# Environment Variables
python-dotenv>=1.0.0

# HTTP Requests
requests>=2.32.0
beautifulsoup4>=4.12.0
lxml>=5.3.0

# Utilities
python-dateutil>=2.9.0
pydantic[email]>=2.10.0
pydantic-settings>=2.7.0
email-validator>=2.0.0

# Testing
pytest>=8.3.0
pytest-asyncio>=0.24.0
httpx>=0.27.0

# Production Server
gunicorn>=21.2.0